Fixes linter error E0611: No name 'RWApiKeysRepository' in module '_helpers' (no-name-in-module)
